+++ VCCTechTest/Models/Product.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double CurrentPrice { get; set; }
+        public Decimal CurrentPrice { get; set; }
 
         //PriceHistory
